Testing added

+++ src/main/java/com/example/AddressBook/model/AddressBook.java
@@ -10,7 +10,7 @@ public class AddressBook implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private String phone;
@@ -18,13 +18,13 @@ public class AddressBook implements Serializable {
 
     public AddressBook(){}
 
-    public  AddressBook(int id, String name, String email, String phone){
+    public  AddressBook(Long id, String name, String email, String phone){
         this.id=id;
         this.name=name;
         this.email=email;
         this.phone=phone;
     }
-    public int getId(){
+    public Long getId(){
         return id;
     }
